# server/openapi.yaml
openapi: 3.0.3
info:
  title: Gas Log Submission API
  version: 1.0.0
  description: API for submitting gas receipts and odometer readings.
paths:
  /submitGas:
    post:
      summary: Submit gas receipt and odometer information
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                vehicleId:
                  type: integer
                  description: ID of the vehicle (required)
                  example: 1
                receiptPhoto:
                  type: string
                  format: binary
                  description: Photo of the gas receipt (required)
                odometerPhoto:
                  type: string
                  format: binary
                  description: Photo of the odometer (required if odometerInputMethod is 'separatephoto')
                odometerReading:
                  type: string
                  description: Manual odometer reading (required if odometerInputMethod is 'manual')
                odometerInputMethod:
                  type: string
                  enum: [separatephoto, onreceipt, manual]
                  description: How the odometer reading is provided
                filledToFull:
                  type: string
                  enum: [yes, no]
                  description: Whether the car was filled to full
                filledLastTime:
                  type: string
                  enum: [yes, no]
                  description: Whether the form was filled last time
              required:
                - vehicleId
                - receiptPhoto
                - odometerInputMethod
                - filledToFull
                - filledLastTime
      responses:
        '200':
          description: Submission successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Form submitted successfully
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing required fields
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Bad Gateway
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Service Unavailable
  /vehicles:
    get:
      summary: Retrieve a list of vehicles
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicles:
                    type: array
                    items:
                      type: object
                      properties:
                        vehicleId:
                          type: integer
                          example: 1
                        year:
                          type: integer
                          example: 2020
                        make:
                          type: string
                          example: Toyota
                        model:
                          type: string
                          example: Camry
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid request
        '500':
            description: Server error
            content:
              application/json:
                schema:
                    type: object
                    properties:
                      error:
                        type: string
                        example: Internal server error
        '502':
            description: Bad Gateway
            content:
              application/json:
                schema:
                    type: object
                    properties:
                      error:
                        type: string
                        example: Bad Gateway
  /health:
    get:
      summary: Check the health of the API
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Bad Gateway
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Service Unavailable